// src/types/delivery.tsx
export type Delivery = {
    deliveryId: number;
    userId: number;
    displayName: string;
    firstName: string;
    secondName: string;
    companyName: string;
    nationalApprovalId: string;
    ownerName: string;
    ownerNationalId: string;
    email: string;
    isActive: boolean;
    totalOrders: number;
    availableBalance: number;
    totalBalance: number;
  };